Introduction of DNS meta records
We want to be able to remove records in templating, for example in order
to have a template set generally, but for a specific domain to set a
different SPF entry we need to remove the templated SPF record (two TXT
SPF records are disallowed). This adds "meta" records which have the
ability to remove templated record entries again.